requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic[email]==2.8.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
redis==5.1.1
aioredis==2.0.1
aio-pika==9.4.3
python-dotenv==1.0.1
aiogram==3.10.0
httpx==0.27.2
Jinja2==3.1.4
python-multipart==0.0.9

